requirements.txt
PyMuPDF>=1.23.0
unstructured>=0.11.0
# Ana AI ve Pydantic kütüphaneleri
google-genai==0.3.0
pydantic==2.6.1
pydantic-settings>=2.0

# Excel Raporlama
pandas==2.2.0
openpyxl==3.1.2

# Sistem Mimarisi & Hata Yönetimi
python-dotenv==1.0.1
tenacity==8.2.3

# Veritabanı ve ORM
SQLAlchemy>=2.0
aiosqlite>=0.19

# Web Arayüzü (Dashboard)
streamlit>=1.31.0
